[WIP] UI: Show fetch options dropdown only when feasible
In getUrl(), the option to use overpass is only given in viewerMode. But the Fetch options dropdown suggests is can be used in both modes.

With this change, the dropdown is hidden unless it actually does something.

WIP => However, this solution feels too hacky to use this way. I had to remove and re-add the fetch controls in order to get a re-render going. Is there a nicer way to do this?