Fix `generateStyleMapByZoom` being out of sync with maxZoom
I received an error `Uncaught TypeError: Cannot read properties of undefined (reading 'offsetMinor')` in my build. But it does not show on https://zlant.github.io/parking-lanes/#21/52.50753/13.46417.

However, this will make sure it does not show up by adding a comment and a bit of help to clarify that the maxZoom needs to be kept in sync between `tileLayers` and `generateStyleMapByZoom()`.