Add host to changeset tags
This provides a link back to the editor which in turn helps readers of the changeset understand what tagging is applied and how.

The host value follows the example of iD (example https://www.openstreetmap.org/changeset/113127802). MapComplete also uses the host key but only the domain as value (example https://www.openstreetmap.org/changeset/113127802).