Hide capacity input unless already specified
We consider it bad practice to map the capacity as number on a way since each split would multiply the capacity unless the mapper is not careful. Instead, we use data processing steps to calculate the capacity in Berlin.

This hides the capacity field unless a values is already present. This way existing data can be edited.

It also uses a clearer placeholder text.